PK int -> Long 으로 변환

+++ src/main/java/com/example/demo/entity/Book.java
@@ -9,7 +9,7 @@ public class Book { // Book --> Table
 
     @Id // PK
     @GeneratedValue(strategy = GenerationType.IDENTITY) // AutoIncrement(AI)
-    private int id; // 일련번호 : 1 2 3 4
+    private Long id; // 일련번호 : 1 2 3 4
     @Column(length = 50, nullable = false) // 길이 지정, null 허용 x
     private String title; // 제목
     private int price; // 가격
